requirements.txt
asgiref==3.2.7
astroid==2.4.2
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
boto3==1.14.1
botocore==1.17.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
dj-database-url==0.5.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
jmespath==0.10.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.1
matplotlib==3.2.1
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.18.4
pandas==1.0.3
parsel==1.6.0
Pillow==7.1.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycparser==2.20
PyDispatcher==2.0.5
PyGetWindow==0.0.8
PyHamcrest==2.0.2
pylint==2.5.3
PyMsgBox==1.0.8
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2019.3
queuelib==1.5.0
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
seaborn==0.10.1
service-identity==18.1.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
w3lib==1.22.0
whitenoise==5.1.0
wrapt==1.12.1
zope.interface==5.1.0
